requirements.txt
# SVIES — Python Dependencies

# ── Core AI / Vision ──
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# ── OCR ──
easyocr>=1.7.0
pytesseract>=0.3.10

# ── AI OCR Verification (Groq) ──
groq>=0.28.0

# ── Geofencing ──
shapely>=2.0.0

# ── Backend API ──
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
slowapi>=0.1.9

# ── Database & Data ──
pandas>=2.0.0
pillow>=10.0.0

# ── Auth & Cloud ──
firebase-admin>=6.0.0
supabase>=2.0.0

# ── Alerts ──
twilio>=8.0.0

# ── PDF Reports ──
reportlab>=4.0.0

# ── Roboflow ──
roboflow>=0.2.0

# ── Config ──
python-dotenv>=1.0.0

# ── Testing ──
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
